tk1 fixes

+++ modules/core/src/MetaObject/core/detail/Allocator.cpp
@@ -4,17 +4,17 @@
 namespace mo
 {
 
-    uint8_t* alignMemory(uint8_t* ptr, const uint64_t elem_size)
+    uint8_t* alignMemory(uint8_t* ptr, const size_t elem_size)
     {
-        return &ptr[alignmentOffset(ptr, size_t(elem_size))];
+        return &ptr[alignmentOffset(ptr, elem_size)];
     }
 
-    const uint8_t* alignMemory(const uint8_t* ptr, const uint64_t elem_size)
+    const uint8_t* alignMemory(const uint8_t* ptr, const size_t elem_size)
     {
-        return &ptr[alignmentOffset(ptr, size_t(elem_size))];
+        return &ptr[alignmentOffset(ptr, elem_size)];
     }
 
-    uint64_t alignmentOffset(const uint8_t* ptr, const uint64_t elem_size)
+    size_t alignmentOffset(const uint8_t* ptr, const size_t elem_size)
     {
         return elem_size - (reinterpret_cast<const size_t>(ptr) % elem_size);
     }
