+++ modules/core/src/MetaObject/core/detail/allocator_policies/Combined.hpp
@@ -12,17 +12,17 @@ namespace mo
         using Ptr = std::shared_ptr<CombinedPolicy>;
         static Ptr create();
 
-        CombinedPolicy(const uint64_t threshold = 1 * 1024 * 512);
+        CombinedPolicy(const size_t threshold = 1 * 1024 * 512);
 
-        uint8_t* allocate(const uint64_t num_bytes, const uint64_t elem_size);
+        uint8_t* allocate(const size_t num_bytes, const size_t elem_size);
 
-        void deallocate(uint8_t* ptr, const uint64_t num_bytes);
+        void deallocate(uint8_t* ptr, const size_t num_bytes);
         void release();
-        void setThreshold(const uint64_t thresh);
-        uint64_t getThreshold() const;
+        void setThreshold(const size_t thresh);
+        size_t getThreshold() const;
 
       private:
-        const uint64_t m_threshold;
+        const size_t m_threshold;
         SmallAllocator m_small_allocator;
         LargeAllocator m_large_allocator;
     };
@@ -37,14 +37,14 @@ namespace mo
     }
 
     template <class SmallAllocator, class LargeAllocator>
-    CombinedPolicy<SmallAllocator, LargeAllocator>::CombinedPolicy(const uint64_t threshold)
+    CombinedPolicy<SmallAllocator, LargeAllocator>::CombinedPolicy(const size_t threshold)
         : m_threshold(threshold)
     {
     }
 
     template <class SmallAllocator, class LargeAllocator>
-    uint8_t* CombinedPolicy<SmallAllocator, LargeAllocator>::allocate(const uint64_t num_bytes,
-                                                                      const uint64_t elem_size)
+    uint8_t* CombinedPolicy<SmallAllocator, LargeAllocator>::allocate(const size_t num_bytes,
+                                                                      const size_t elem_size)
     {
         if (num_bytes > m_threshold)
         {
@@ -57,7 +57,7 @@ namespace mo
     }
 
     template <class SmallAllocator, class LargeAllocator>
-    void CombinedPolicy<SmallAllocator, LargeAllocator>::deallocate(uint8_t* ptr, const uint64_t num_bytes)
+    void CombinedPolicy<SmallAllocator, LargeAllocator>::deallocate(uint8_t* ptr, const size_t num_bytes)
     {
         if (num_bytes > m_threshold)
         {
@@ -77,13 +77,13 @@ namespace mo
     }
 
     template <class SmallAllocator, class LargeAllocator>
-    void CombinedPolicy<SmallAllocator, LargeAllocator>::setThreshold(const uint64_t thresh)
+    void CombinedPolicy<SmallAllocator, LargeAllocator>::setThreshold(const size_t thresh)
     {
         m_threshold = thresh;
     }
 
     template <class SmallAllocator, class LargeAllocator>
-    uint64_t CombinedPolicy<SmallAllocator, LargeAllocator>::getThreshold() const
+    size_t CombinedPolicy<SmallAllocator, LargeAllocator>::getThreshold() const
     {
         return m_threshold;
     }
