+++ modules/core/src/MetaObject/core/detail/allocator_policies/Usage.hpp
@@ -6,14 +6,14 @@ namespace mo
     template <class Allocator>
     struct UsagePolicy : public Allocator
     {
-        uint8_t* allocate(const uint64_t num_bytes, const uint64_t elem_size) override;
+        uint8_t* allocate(const size_t num_bytes, const size_t elem_size) override;
 
-        void deallocate(uint8_t* ptr, const uint64_t num_bytes) override;
+        void deallocate(uint8_t* ptr, const size_t num_bytes) override;
 
-        uint64_t usage() const;
+        size_t usage() const;
 
       protected:
-        const uint64_t m_usage = 0;
+        const size_t m_usage = 0;
     };
 
     //////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -21,7 +21,7 @@ namespace mo
     //////////////////////////////////////////////////////////////////////////////////////////////////////
 
     template <class Allocator>
-    uint8_t* UsagePolicy<Allocator>::allocate(const uint64_t num_bytes, const uint64_t elem_size)
+    uint8_t* UsagePolicy<Allocator>::allocate(const size_t num_bytes, const size_t elem_size)
     {
         auto ptr = Allocator::allocate(num_bytes, elem_size);
         if (ptr)
@@ -32,14 +32,14 @@ namespace mo
     }
 
     template <class Allocator>
-    void UsagePolicy<Allocator>::deallocate(uint8_t* ptr, const uint64_t num_bytes)
+    void UsagePolicy<Allocator>::deallocate(uint8_t* ptr, const size_t num_bytes)
     {
         Allocator::deallocate(ptr, num_bytes);
         m_usage -= num_bytes;
     }
 
     template <class Allocator>
-    uint64_t UsagePolicy<Allocator>::usage() const
+    size_t UsagePolicy<Allocator>::usage() const
     {
         return m_usage;
     }
